refactor filter class names
